# docs/swagger.yaml
definitions:
  models.People:
    properties:
      address:
        type: string
      id:
        type: integer
      name:
        type: string
      passportNumber:
        type: string
      patronymic:
        type: string
      surname:
        type: string
    type: object
  models.WorkLog:
    properties:
      duration:
        type: string
      end_time:
        example: "2023-01-01T17:04:05Z"
        type: string
      id:
        type: integer
      start_time:
        example: "2023-01-01T15:04:05Z"
        type: string
      task_id:
        example: 100
        type: integer
      user_id:
        example: 1
        type: integer
    type: object
info:
  contact: {}
paths:
  /createuser:
    post:
      consumes:
      - application/json
      description: Create a new user with the input payload
      parameters:
      - description: User to create
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.People'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.People'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create a new user
      tags:
      - Users
  /deleteuser:
    delete:
      consumes:
      - application/json
      description: Delete a user with the provided details
      parameters:
      - description: User to delete
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.People'
      produces:
      - application/json
      responses:
        "200":
          description: User deleted successfully
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete a user
      tags:
      - Users
  /startwork:
    post:
      consumes:
      - application/json
      description: Start work on a task for a user
      parameters:
      - description: Work start input
        in: body
        name: input
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Work started successfully
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Start work
      tags:
      - Work
  /stopwork:
    post:
      consumes:
      - application/json
      description: Stop work on a task for a user
      parameters:
      - description: Work stop input
        in: body
        name: input
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Work stopped successfully
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Stop work
      tags:
      - Work
  /updateuser:
    put:
      consumes:
      - application/json
      description: Update user details with the input payload
      parameters:
      - description: User to update
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.People'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.People'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update a user
      tags:
      - Users
  /users:
    get:
      consumes:
      - application/json
      description: Get list of users based on filter parameters
      parameters:
      - description: User filter
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.People'
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Number of items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.People'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get list of users
      tags:
      - Users
  /worklogs:
    get:
      description: Get work logs for a user within a specified date range
      parameters:
      - description: User ID
        in: query
        name: userID
        required: true
        type: integer
      - description: Start date
        format: date-time
        in: query
        name: start
        required: true
        type: string
      - description: End date
        format: date-time
        in: query
        name: end
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.WorkLog'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get work logs
      tags:
      - WorkLogs
swagger: "2.0"